Use octagonal tube for game world and make player snap to the edges of the tunnel

+++ src/Game/Player.h
@@ -31,8 +31,8 @@ class Player : public GameObject {
   bool lastFireState;
 
   //Directional inputs
-  float netX;
-  float netY;
+  int netX;
+  int netY;
 
   float fireRate;
   float nextFireTime;
